add CommandScheduleTest, improve test base
